Refactor Dataset object
The dataset object now tracks its processing parameters (connection_range and time_step_size) in a config.json file. Before reprocessing it now checks whether the parameters have changed. If not the processing step will be skipped.

With the new implementation you no longer need to provide a skip_processing parameter since the reprocessing check is performed automatically.

The implementation is based on the Dataset class of the pytorch_geometric framework. However, there are some changes since the original class does not allow parameter passing in the processing step.